requirements.txt
azure-ai-documentintelligence
openai
python-dotenv
streamlit